Clamp shooting and bullet tag
When shooting at cursor, gun rotation and bullet velocity clamped so can
only shoot between straight up and straight down (and anywhere between
in front of character). Bullets have a "shooter tag" that needs to be
set when firing a bullet. This should be set to gun's parent's tag. the
bullet checks this when colliding to determine whether deals damage to
collided.