Player Sprites
Right, this is a bit more complicated. This is how complicated it gets
to have three substates (aiming up, middle, and down) for each state of
action (idle, shooting, and jumping), and THEN having a SHOOTING version
of each of those. It might be unnecessarily complicated, but it's not
TOO difficult to work on my end. As long as I can figure out a way to
easily generate all the sprite sheets it should be fine. Right now
that's a serious pain in the ass, however.